[libclang/python][ci] Add release Clang Python Bindings CI workflow
This migrates the CI workflow from https://github.com/trolldbois/python-clang/blob/master/.github/workflows/release.yml and adapts it to work within the structure of the LLVM monorepo.

It builds upon the Python packaging files added in #125806, and should be merged after that one.

The CI workflow is set up to use two jobs:
- the first just builds the package and saves it as an artifact and will work for testing in forks, PRs, and untagged commits
- the second checks for release manager permissions and uses trusted publishing to upload the package to PyPI

Signed-off-by: Ryan Mast <mast.ryan@gmail.com>